+++ client/package.json
@@ -13,6 +13,7 @@
     "prepare": "cd .. && husky client/.husky"
   },
   "dependencies": {
+    "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.3.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
