+++ client/src/App.jsx
@@ -5,14 +5,12 @@ import ProjectsPage from './pages/ProjectsPage';
 import NewsPage from './pages/NewsPage';
 import Footer from './layout/Footer';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import EventsSection from './layout/EventsSection';
 
 function App() {
   return (
     <BrowserRouter>
       <>
         <Navbar />
-        <EventsSection />
         <Routes>
           <Route path="/" element={<Homepage />} />
           <Route path="/events" element={<EventsPage />} />
