--- client/src/layout/EventsSection.module.css
@@ -1,116 +1,0 @@
-/* EventsSection.module.css */
-
-.eventsHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.eventsHeading {
-  font-size: 2.5rem;
-  font-weight: bold;
-  font-family: 'Nacelle-Black', sans-serif;
-  margin-bottom: 20px;
-  margin-top: 40px;
-}
-
-.allEventsButton {
-  font-size: 1rem;
-  background-color: #e1f264;
-  color: #000;
-  border: none;
-  padding: 10px 20px;
-  cursor: pointer;
-  border-radius: 5px;
-  transition: background-color 0.3s;
-}
-
-.allEventsButton:hover {
-  background-color: #cfd554;
-}
-
-.events {
-  display: flex;
-  justify-content: center;
-}
-
-.eventsItem {
-  background-image: url('../img/eventspic.jpg');
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  width: 300px; /* Adjust this value as needed for card width */
-  padding: 20px;
-  border-radius: 15px;
-  margin: 0 15px; /* Increased margin for gaps between cards */
-  position: relative;
-  transition: transform 0.3s;
-}
-
-.eventsItem:hover {
-  transform: translateY(-10px);
-}
-
-.eventContent {
-  background: rgba(0, 0, 0, 0.6); /* Add a dark overlay */
-  color: #fff;
-  padding: 20px;
-  border-radius: 10px;
-}
-
-.eventTitle {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin-bottom: 10px;
-}
-
-.eventDetails {
-  font-size: 1rem;
-  line-height: 1.5;
-}
-
-.registerButton {
-  font-size: 1rem;
-  background-color: #e1f264;
-  color: #000;
-  border: none;
-  padding: 10px 15px;
-  cursor: pointer;
-  border-radius: 5px;
-  margin-top: 10px;
-  transition: background-color 0.3s;
-}
-
-.registerButton:hover {
-  background-color: #cfd554;
-}
-
-/* Responsive styles */
-@media (max-width: 1000px) {
-  .eventsItem {
-    width: 250px;
-    margin: 0 10px; /* Adjust margins for smaller screens */
-  }
-}
-
-@media (max-width: 600px) {
-  .eventsContainer {
-    padding: 10px;
-  }
-
-  .eventsItem {
-    width: 100%;
-    margin: 10px 0; /* Stacking cards vertically */
-  }
-
-  .eventsHeading {
-    font-size: 2rem;
-  }
-}
-
-@media (max-width: 300px) {
-  .eventsHeading {
-    font-size: 1.8rem;
-  }
-}
